requirements.txt
attrdict==2.0.0
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==2.0.0
eth-utils==1.4.0
Flask==1.0.2
hexbytes==0.1.0
idna==2.8
kiwisolver==1.0.1
lesscpy==0.13.0
lru-dict==1.1.6
py-solc==3.2.0
PyYAML==3.13
requests==2.21.0
traitlets==4.3.2
urllib3==1.24.1
web3==4.8.2
webencodings==0.5.1
websockets==6.0
Werkzeug==0.14.1
zxing==0.9.3
